+++ server/src/const/boj-results.ts
