+++ server/src/room-user/room-user.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import RoomUser from 'src/entities/roomUser.entity';
-import { RoomUserService } from './room.user.service';
+import RoomUser from 'src/entities/room-user.entity';
+import { RoomUserService } from './room-user.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([RoomUser])],
