+++ server/src/room/room.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import Room from 'src/entities/room.entity';
-import { RoomUserModule } from 'src/roomUser/room.user.module';
+import { RoomUserModule } from 'src/room-user/room-user.module';
 import { UserModule } from 'src/user/user.module';
 import { RoomController } from './room.controller';
 import { RoomService } from './room.service';
