+++ server/src/submission/submission.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import Submission from 'src/entities/submission.entity';
 import { ProblemModule } from 'src/problem/problem.module';
-import { RoomUserModule } from 'src/roomUser/room.user.module';
+import { RoomUserModule } from 'src/room-user/room-user.module';
 import { UserModule } from 'src/user/user.module';
 import { SubmissionController } from './submission.controller';
 import { SubmissionService } from './submission.service';
