+++ server/src/types/message-interface.ts
